src/main/python/requirements.txt
nltk==3.5
pyspark
kafka-python
psycopg2
